+++ src/app/library/detail/[id]/_components/WritableUserModal.tsx
@@ -5,11 +5,11 @@ import { addDays, format } from 'date-fns';
 import CreateStoryContentModal from './CreateContentModal';
 import { WritableUserModalProps } from './type';
 import { useEffect, useState } from 'react';
-import useGetApproveUser from '@/hooks/api/content-approvals/useGetApproveUser';
-import useApproveContent from '@/hooks/api/content-approvals/useApproveContent';
+import useGetApproveUser from '@/hooks/api/supabase/content-approvals/useGetApproveUser';
+import useApproveContent from '@/hooks/api/supabase/content-approvals/useApproveContent';
 import { CircleCheck } from 'lucide-react';
-import useGetLastContent from '@/hooks/api/contents/useGetLastContent';
-import useGetSocialParticipantsByDb from '@/hooks/api/useGetSocialParticipantsByDb';
+import useGetLastContent from '@/hooks/api/supabase/contents/useGetLastContent';
+import useGetSocialParticipantsByDb from '@/hooks/api/supabase/useGetSocialParticipantsByDb';
 import { useStoryModal } from '@/providers/StoryWriteOrApproveModalProviders';
 
 const WritableUserModal = ({
