+++ src/hooks/api/supabase/story-collaborators/useGetSocialDetail.ts
